♻️🏢 Extract loader
