+++ apps/admin/src/routes/_layout.show.exhibitors.$id.edit.stand-configuration/form.ts
@@ -5,7 +5,8 @@ import { useForm as useFormBase } from "@conform-to/react";
 import { getZodConstraint, parseWithZod } from "@conform-to/zod";
 import { useFetcher, useLoaderData } from "@remix-run/react";
 import { ActionSchema, DividerType } from "./action";
-import type { action, loader } from "./route";
+import type { loader } from "./loader.server.js";
+import type { action } from "./route";
 
 export function useFormRoot() {
   const { exhibitor } = useLoaderData<typeof loader>();
