+++ apps/admin/src/routes/_layout.show.exhibitors.$id.edit.stand-configuration/route.tsx
@@ -10,67 +10,20 @@ import type { SubmissionResult } from "@conform-to/react";
 import { getFormProps } from "@conform-to/react";
 import { parseWithZod } from "@conform-to/zod";
 import { UserGroup } from "@prisma/client";
-import type { ActionFunctionArgs, LoaderFunctionArgs } from "@remix-run/node";
+import type { ActionFunctionArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/react";
-import { promiseHash } from "remix-utils/promise";
 import type { MergeExclusive } from "type-fest";
 import { ActionSchema } from "./action";
 import { FieldsetConfiguration } from "./fieldset-configuration";
 import { FieldsetPriceDetails } from "./fieldset-price-details";
 import { FieldsetStatus } from "./fieldset-status";
 import { FormProvider, useFormRoot } from "./form";
+import type { loader } from "./loader.server.js";
 
-export async function loader({ request, params }: LoaderFunctionArgs) {
-  const currentUser = await db.currentUser.get(request, {
-    select: { groups: true },
-  });
+export { loader } from "./loader.server.js";
 
-  assertCurrentUserHasGroups(currentUser, [
-    UserGroup.ADMIN,
-    UserGroup.SHOW_ORGANIZER,
-  ]);
-
-  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
-
-  const { exhibitor, standSizes, dividerTypes } = await promiseHash({
-    exhibitor: db.show.exhibitor.findUnique(routeParams.id, {
-      select: {
-        name: true,
-        category: true,
-        chairCount: true,
-        dividerCount: true,
-        dividerType: { select: { id: true } },
-        hasElectricalConnection: true,
-        hasTableCloths: true,
-        installationDay: true,
-        peopleCount: true,
-        size: { select: { id: true } },
-        standConfigurationStatus: true,
-        standConfigurationStatusMessage: true,
-        tableCount: true,
-      },
-    }),
-
-    standSizes: db.show.standSize.findManyWithAvailability({
-      select: {
-        id: true,
-        label: true,
-        priceForAssociations: true,
-        priceForServices: true,
-        priceForShops: true,
-      },
-    }),
-
-    dividerTypes: db.show.dividerType.findManyWithAvailability({
-      select: { id: true, label: true },
-    }),
-  });
-
-  return json({ exhibitor, standSizes, dividerTypes });
-}
-
-const RouteParamsSchema = zu.object({
+export const RouteParamsSchema = zu.object({
   id: zu.string().uuid(),
 });
